feat: Extract `Errors` to aya-multierror crate
Previously, the same pattern for aggregating multiple errors with
formatted Display output was duplicated in xtask/src/lib.rs and
`test-distro/src/init.rs`.

Create a new aya-multierror crate with the `Errors<E>` type that can be
re-used.
